Final Help Page Update
Added functionality to replace the form with "Help Form Submitted" text when the user submits the form. Added functionality, if the form fails to submit, to show an error message below the submit button prompting the user to try submitting again. Removed help heading at the top of the page.